requirements.txt
anthropic>=0.34.0
python-dotenv>=1.0.0

